feature/VIH-6525 : Fix defence advocate null
