+++ VideoWeb/VideoWeb.Common/Caching/ConferenceCacheMapper.cs
@@ -42,19 +42,4 @@ namespace VideoWeb.Common.Caching
             return conference;
         }
     }
-
-    public static class EndpointCacheMapper
-    {
-        public static Endpoint MapEndpointToCacheModel(EndpointResponse endpointResponse)
-        {
-            return new Endpoint
-            {
-                Id = endpointResponse.Id,
-                DisplayName = endpointResponse.Display_name,
-                EndpointStatus =
-                    (EndpointStatus) Enum.Parse(typeof(EndpointStatus), endpointResponse.Status.ToString()),
-                DefenceAdvocateUsername = endpointResponse.Defence_advocate.ToLower().Trim()
-            };
-        }
-    }
 }
