+++ VideoWeb/VideoWeb.Common/Caching/EndpointCacheMapper.cs
@@ -1,0 +1,21 @@
+using System;
+using VideoWeb.Common.Models;
+using VideoWeb.Services.Video;
+
+namespace VideoWeb.Common.Caching
+{
+    public static class EndpointCacheMapper
+    {
+        public static Endpoint MapEndpointToCacheModel(EndpointResponse endpointResponse)
+        {
+            return new Endpoint
+            {
+                Id = endpointResponse.Id,
+                DisplayName = endpointResponse.Display_name,
+                EndpointStatus =
+                    (EndpointStatus) Enum.Parse(typeof(EndpointStatus), endpointResponse.Status.ToString()),
+                DefenceAdvocateUsername = endpointResponse.Defence_advocate?.ToLower().Trim()
+            };
+        }
+    }
+}
